+++ Millicast SDK Sample App in Swift/Utils.swift
@@ -3,6 +3,7 @@
 //  Millicast SDK Sample App in Swift
 //
 
+import AVFAudio
 import Foundation
 
 /**
@@ -10,7 +11,7 @@ import Foundation
  */
 class Utils {
     /**
-     Get a String representing the specified CredentialSource.
+     * Gets a String representing the specified CredentialSource.
      */
     public static func getCredStr(creds: CredentialSource) -> String {
         let str = "Account ID: \(creds.getAccountId())\n\t" +
@@ -24,8 +25,174 @@ class Utils {
     }
 
     /**
-     Get a String value from UserDefaults, if available.
-     If not, return the specified defaultValue.
+     * Configures the AVAudioSession to use:
+     * - Category playAndRecord for iOS, or playback for tvOS, with the mixWithOthers option.
+     *  - This will allow:
+     *   - Recording on iOS.
+     *   - Audio playback in the backgroud, and even when the screen is locked.
+     *   - Mixing with audio from other apps that also allow mixing.
+     * - Mode videoChat, which will optimise the audio for voice and allow Bluetooth Hands-Free Profile (HFP) device as input and output.
+     * - Audio will default to the device's speakers when no other audio route is connected.
+     * For full control of the AVAudioSession, this method should be called:
+     * - When the Subscriber's audioTrack is rendered, for e.g. at MillicastManager.subRenderAudio(track: MCAudioTrack?).
+     * - When the AVAudioSession route changes, for e.g. at MillicastSA.routeChangeHandler(notification: Notification).
+     */
+    public static func configureAudioSession() {
+        let logTag = "[Configure][Audio][Session] "
+        let session = AVAudioSession.sharedInstance()
+        print(logTag + "Now: " + Utils.audioSessionStr(session: session))
+        do {
+            #if os(iOS)
+            try session.setCategory(AVAudioSession.Category.playAndRecord, mode: .videoChat, options: [.mixWithOthers])
+            #else
+            try session.setCategory(AVAudioSession.Category.playback, options: [.mixWithOthers])
+            #endif
+            try session.setActive(true)
+        } catch {
+            print(logTag + "Failed! Error: \(error)")
+            return
+        }
+        print(logTag + "OK. " + Utils.audioSessionStr(session: session))
+    }
+
+    /**
+     * Sets the AVAudioSession to active or not active as indicated by the parameter.
+     */
+    public static func setAudioSession(active: Bool) {
+        #if os(iOS)
+        let logTag = "[Configure][Audio][Session][Active] "
+        let session = AVAudioSession.sharedInstance()
+        do {
+            if active {
+                try session.setActive(true)
+            } else {
+                try session.setActive(false, options: AVAudioSession.SetActiveOptions.notifyOthersOnDeactivation)
+            }
+        } catch {
+            print(logTag + "Failed! Error: \(error)")
+            return
+        }
+        print(logTag + "OK. SetActive \(active).")
+        #endif
+    }
+
+    /**
+     * Gets a String representing characteristics of the current AVAudioSession.
+     */
+    public static func audioSessionStr(session: AVAudioSession) -> String {
+        let cat = String(describing: session.category.rawValue)
+        let catOpts = audioCatOptStr(value: session.categoryOptions.rawValue)
+        let mode = String(describing: session.mode.rawValue)
+        let gain = session.inputGain
+        let vol = session.outputVolume
+        let channelNumIn = session.inputNumberOfChannels
+        let channelNumOut = session.outputNumberOfChannels
+        let portsIn = inputAudioStr(route: session.currentRoute)
+        let portsOut = outputAudioStr(route: session.currentRoute)
+        var oriIn = "-"
+        #if os(iOS)
+        oriIn = String(describing: session.inputOrientation.rawValue)
+        #endif
+
+        let str = "AVAudioSession: Category:\(cat) options:\(catOpts) Mode:\(mode) Gain:\(gain) Vol:\(vol) Channels in:\(channelNumIn) out:\(channelNumOut) Ports in:\(portsIn) out:\(portsOut) orientation:\(oriIn)."
+        return str
+    }
+
+    /**
+     * Gets a String representing the specified AVAudioSession.CategoryOptions.
+     */
+    public static func audioCatOptStr(value: UInt) -> String {
+        var opt = AVAudioSession.CategoryOptions(rawValue: value)
+        var str = ""
+
+        addOptToStr(catOpt: .mixWithOthers, desc: "mixWithOthers")
+        addOptToStr(catOpt: .duckOthers, desc: "duckOthers")
+        #if os(iOS)
+        addOptToStr(catOpt: .allowBluetooth, desc: "allowBluetooth")
+        addOptToStr(catOpt: .defaultToSpeaker, desc: "defaultToSpeaker")
+        #endif
+        addOptToStr(catOpt: .interruptSpokenAudioAndMixWithOthers, desc: "interruptSpokenAudioAndMixWithOthers")
+        addOptToStr(catOpt: .allowBluetoothA2DP, desc: "allowBluetoothA2DP")
+        addOptToStr(catOpt: .allowAirPlay, desc: "allowAirPlay")
+        #if os(iOS)
+        addOptToStr(catOpt: .overrideMutedMicrophoneInterruption, desc: "overrideMutedMicrophoneInterruption")
+        #endif
+
+        if str.isEmpty {
+            str = "-"
+        } else {
+            str = String(str.dropFirst(1))
+        }
+        str = "[\(str)]"
+
+        func addOptToStr(catOpt: AVAudioSession.CategoryOptions, desc: String) {
+            if !opt.contains(catOpt) {
+                return
+            }
+            str += "," + desc
+        }
+
+        return str
+    }
+
+    /**
+     Logs the current and previous audio output device.
+     */
+    public static func audioOutputLog(userInfo: [AnyHashable: Any]) -> String {
+        var log = ""
+        // Get current
+        let session = AVAudioSession.sharedInstance()
+        let routeCur = session.currentRoute
+        let routeCurStr = outputAudioStr(route: routeCur)
+        log += "Cur: \(routeCurStr), Old: "
+
+        // Get previous
+        guard let routePrev =
+            userInfo[AVAudioSessionRouteChangePreviousRouteKey] as? AVAudioSessionRouteDescription
+        else {
+            log += "None!"
+            return log
+        }
+
+        let routePrevStr = outputAudioStr(route: routePrev)
+        log += "\(routePrevStr)"
+
+        return log
+    }
+
+    /**
+     * Gets a String representation of the audio input port(s) in the given AVAudioSessionRouteDescription.
+     */
+    public static func inputAudioStr(route routeDescription: AVAudioSessionRouteDescription) -> String {
+        var ports = portStr(ports: routeDescription.inputs)
+        return ports
+    }
+
+    /**
+     * Gets a String representation of the audio output port(s) in the given AVAudioSessionRouteDescription.
+     */
+    public static func outputAudioStr(route routeDescription: AVAudioSessionRouteDescription) -> String {
+        var ports = portStr(ports: routeDescription.outputs)
+        return ports
+    }
+
+    /**
+     * Gets a String representation of the given list of AVAudioSessionPortDescription.
+     */
+    public static func portStr(ports: [AVAudioSessionPortDescription]) -> String {
+        var portStr = ""
+        for port in ports {
+            portStr += "\(port.portType.rawValue),"
+        }
+        if portStr != "" {
+            portStr.removeLast()
+        }
+        return portStr
+    }
+
+    /**
+     * Gets a String value from UserDefaults, if available.
+     * If not, return the specified defaultValue.
      */
     public static func getValue(tag: String, key: String, defaultValue: String) -> String {
         var value: String
@@ -44,8 +211,8 @@ class Utils {
     }
 
     /**
-     Get an integer value from UserDefaults, if available.
-     If not, return the specified defaultValue.
+     * Gets an integer value from UserDefaults, if available.
+     * If not, return the specified defaultValue.
      */
     public static func getValue(tag: String, key: String, defaultValue: Int) -> Int {
         var value: Int